+++ saleor/graphql/order/filters.py
@@ -12,7 +12,7 @@ from ...core.postgres import FlatConcat
 from ...giftcard import GiftCardEvents
 from ...giftcard.models import GiftCardEvent
 from ...invoice.models import Invoice
-from ...order.models import Fulfillment, Order, OrderLine
+from ...order.models import Fulfillment, Order, OrderEvent, OrderLine
 from ...order.search import search_orders
 from ...payment import ChargeStatus
 from ...product.models import ProductVariant
@@ -27,12 +27,13 @@ from ..core.filters import (
     ObjectTypeFilter,
     ObjectTypeWhereFilter,
     OperationObjectTypeWhereFilter,
-    WhereFilterSet,
 )
+from ..core.filters.where_filters import MetadataWhereBase, filter_where_metadata
 from ..core.filters.where_input import (
     FilterInputDescriptions,
     GlobalIDFilterInput,
     IntFilterInput,
+    MetadataFilterInput,
     PriceFilterInput,
     StringFilterInput,
     UUIDFilterInput,
@@ -62,6 +63,7 @@ from .enums import (
     FulfillmentStatusEnum,
     OrderAuthorizeStatusEnum,
     OrderChargeStatusEnum,
+    OrderEventsEnum,
     OrderStatusEnum,
     OrderStatusFilter,
 )
@@ -274,6 +276,23 @@ def filter_has_fulfillments(qs, value):
     return qs.filter(~Exists(fulfillments))
 
 
+def filter_fulfillments(qs, value):
+    if value is None:
+        return qs.none()
+    fulfillment_qs = None
+    if status_value := value.get("status"):
+        fulfillment_qs = filter_where_by_value_field(
+            Fulfillment.objects.using(qs.db), "status", status_value
+        )
+    if metadata_value := value.get("metadata"):
+        fulfillment_qs = filter_where_metadata(
+            fulfillment_qs or Fulfillment.objects.using(qs.db), None, metadata_value
+        )
+    if fulfillment_qs is not None:
+        return qs.filter(Exists(fulfillment_qs.filter(order_id=OuterRef("id"))))
+    return qs.none()
+
+
 class DraftOrderFilter(MetadataFilterBase):
     customer = django_filters.CharFilter(method=filter_customer)
     created = ObjectTypeFilter(input_class=DateRangeInput, method=filter_created_range)
@@ -398,14 +417,46 @@ class FulfillmentFilterInput(BaseInputObjectType):
     status = FulfillmentStatusEnumFilterInput(
         description="Filter by fulfillment status."
     )
+    metadata = MetadataFilterInput(description="Filter by metadata fields.")
 
     class Meta:
         doc_category = DOC_CATEGORY_ORDERS
-        description = "Filter input for fulfillments."
+        description = "Filter input for order fulfillments data."
 
 
-# TODO: metadata filter will be added later
-class OrderWhere(WhereFilterSet):
+class LinesFilterInput(BaseInputObjectType):
+    metadata = MetadataFilterInput(
+        description="Filter by metadata fields of order lines."
+    )
+
+    class Meta:
+        doc_category = DOC_CATEGORY_ORDERS
+        description = "Filter input for order lines data."
+
+
+class OrderEventTypeEnumFilterInput(BaseInputObjectType):
+    eq = OrderEventsEnum(description=FilterInputDescriptions.EQ, required=False)
+    one_of = NonNullList(
+        OrderEventsEnum,
+        description=FilterInputDescriptions.ONE_OF,
+        required=False,
+    )
+
+
+class OrderEventFilterInput(BaseInputObjectType):
+    date = DateTimeRangeInput(
+        description="Filter order events by date.",
+    )
+    type = OrderEventTypeEnumFilterInput(
+        description="Filter order events by type.",
+    )
+
+    class Meta:
+        doc_category = DOC_CATEGORY_ORDERS
+        description = "Filter input for order events data."
+
+
+class OrderWhere(MetadataWhereBase):
     ids = GlobalIDMultipleChoiceWhereFilter(method=filter_by_ids("Order"))
     number = OperationObjectTypeWhereFilter(
         input_class=IntFilterInput,
@@ -501,6 +552,11 @@ class OrderWhere(WhereFilterSet):
         method="filter_fulfillments",
         help_text="Filter by fulfillment data associated with the order.",
     )
+    lines = ObjectTypeWhereFilter(
+        input_class=LinesFilterInput,
+        method="filter_lines",
+        help_text="Filter by metadata fields of order lines.",
+    )
     lines_count = OperationObjectTypeWhereFilter(
         input_class=IntFilterInput,
         method="filter_lines_count",
@@ -515,6 +571,16 @@ class OrderWhere(WhereFilterSet):
         input_class=PriceFilterInput,
         method="filter_total_net",
         help_text="Filter by total net amount of the order.",
+    )
+    product_type_id = OperationObjectTypeWhereFilter(
+        input_class=GlobalIDFilterInput,
+        method="filter_product_type_id",
+        help_text="Filter by the product type of related order lines.",
+    )
+    events = ObjectTypeWhereFilter(
+        input_class=OrderEventFilterInput,
+        method="filter_events",
+        help_text="Filter by order events.",
     )
 
     @staticmethod
@@ -612,13 +678,17 @@ class OrderWhere(WhereFilterSet):
 
     @staticmethod
     def filter_fulfillments(qs, _, value):
-        if value is None:
-            return qs.none()
-        if filter_value := value.get("status"):
-            fulfillments = filter_where_by_value_field(
-                Fulfillment.objects.using(qs.db), "status", filter_value
+        return filter_fulfillments(qs, value)
+
+    @staticmethod
+    def filter_lines(qs, _, value):
+        if not value:
+            return qs
+        if metadata_value := value.get("metadata"):
+            lines_qs = filter_where_metadata(
+                OrderLine.objects.using(qs.db), None, metadata_value
             )
-            return qs.filter(Exists(fulfillments.filter(order_id=OuterRef("id"))))
+            return qs.filter(Exists(lines_qs.filter(order_id=OuterRef("id"))))
         return qs.none()
 
     @staticmethod
@@ -632,6 +702,34 @@ class OrderWhere(WhereFilterSet):
     @staticmethod
     def filter_total_net(qs, _, value):
         return filter_where_by_price_field(qs, "total_net_amount", value)
+
+    @staticmethod
+    def filter_product_type_id(qs, _, value):
+        if not value:
+            return qs
+
+        line_qs = filter_where_by_id_field(
+            OrderLine.objects.using(qs.db), "product_type_id", value, "ProductType"
+        )
+        return qs.filter(Exists(line_qs.filter(order_id=OuterRef("id"))))
+
+    @staticmethod
+    def filter_events(qs, _, value):
+        if not value:
+            return qs.none()
+        if not {"date", "type"}.intersection(value.keys()):
+            return qs.none()
+        if filter_value := value.get("date"):
+            events = filter_where_by_range_field(
+                OrderEvent.objects.using(qs.db), "date", filter_value
+            )
+            qs = qs.filter(Exists(events.filter(order_id=OuterRef("id"))))
+        if filter_value := value.get("type"):
+            events = filter_where_by_value_field(
+                OrderEvent.objects.using(qs.db), "type", filter_value
+            )
+            qs = qs.filter(Exists(events.filter(order_id=OuterRef("id"))))
+        return qs
 
 
 class OrderWhereInput(WhereInputObjectType):